Create CompilerProblem.py
CompilerProblem.py is a Python script that simulates a compiler problem. The script defines two classes, Program and CompilerProblem.

The Program class represents a program with 3 attributes: a name of the program, a length of the program and a frequency of the program call. The __init__ method initializes these attributes when a new Program object is created. The __str__ method returns a string representation of the object in the format "[name, length, frequency]".

The CompilerProblem class has one attribute, num_programs, which defaults to 10 if no value is provided. In the __init__ method, it creates a list of Program objects with random length and frequency values, and prints out the list of programs. The getOptimalOrder method sorts the list of programs by the ratio of frequency to length (using the mergeSort method) and then calculates the total time required to run all of the programs in that order by multiplying the length of each program by its frequency, and adding that to the total time so far. The optimal order of the programs and the total time are printed. The mergeSort method is a recursive method that sorts the program list based on the ratio of frequency to length. It uses the merge sort algorithm to sort the list.