Create AeroplaneProblem.py
This code is a solution to the problem of finding the cheapest path in a matrix where each cell has a cost to move left or down. It consists of two classes, Node and AeroplaneProblem.

The Node class represents a cell in the matrix, with the following properties:

x: the cost to move left from this cell
y: the cost to move down from this cell
price: the best price from the start (0,0) to this cell
priceFromTheEnd: the best price from the end (m,n) to this cell
The AeroplaneProblem class represents the problem and has several methods to solve different variations of the problem:

getBestPrice(): builds a matrix that contains the price to get to each point from the start (0,0) and the number of shortest paths to each point. Complexity: O(n*m) where n is the number of rows and m is the number of columns in the matrix.

getCheapestPrice(): returns the cheapest price from the start to the end.

printNodes(): prints the matrix of nodes

printPrices(): prints the matrix of prices

getOneCheapestPath(): calculates one of the cheapest paths by induction. Complexity: O(n+m) where n and m are the number of rows and columns in the matrix.

getAllCheapestPaths(): calculates all the cheapest paths from the start to the end. Complexity: O(2^(n+m)) where n and m are the number of rows and columns in the matrix.

getCheapestPathFromEnd(): calculates the cheapest path from the end to the start. Complexity: O(n*m) where n and m are the number of rows and columns in the matrix.